% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MMT.R
\name{get_MMT}
\alias{get_MMT}
\title{Get the MMT}
\usage{
get_MMT(crosspred)
}
\arguments{
\item{crosspred}{An object of class "\code{crosspred}" or "\code{crossreduce}".}
}
\value{
The numeric value of the exposure corresponding to the minimum value of the response.
}
\description{
When making predictions with distributed lag nonlinear models, this function
retrieves the value of the exposure that corresponds to the minimum value of
the response. These models are commonly used to predict mortality as a function
of temperature, for which this function returns the minimum mortality temperature
(MMT).
}
\examples{
# The NYC airquality sample dataset
help(airquality)

# Arguments to specify cross-basis
arg_var <- list(fun = "ns", knots = quantile(airquality$Temp, c(.1, .75, .9)))
arg_lag <- list(fun = "ns", knots = logknots(7, nk = 3))

# Cross-basis and model fit
cb <- crossbasis(airquality$Temp, lag = 7, argvar = arg_var, arglag = arg_lag)
fit <- glm(Ozone ~ cb, family = "quasipoisson", data = airquality)

# Prediction with default centering
pred <- crosspred(cb, fit)
plot(pred, "overall")

# Prediction centering at the MMT
pred_centered <- crosspred(cb, fit, cen = get_MMT(pred))
plot(pred_centered, "overall")
}
